// src/sections/form-view/SponsorRegistration/enums.ts
export enum SponsorshipCategory {
  EXHIBITION = "EXHIBITION",
  STATE_PARTNER = "STATE_PARTNER",
  COUNTRY_PARTNER = "COUNTRY_PARTNER",
  KNOWLEDGE_SESSIONS = "KNOWLEDGE_SESSION",
}

export enum ExhibitionPackage {
  PLATINUM = "PLATINUM_PARTNER",
  GOLD = "GOLD_PARTNER",
  SILVER = "SILVER_PARTNER",
  TEXTILE_TECHNOLOGY = "TEXTILE_TECHNOLOGY_PARTNER",
  TEXTILE_INNOVATION = "TEXTILE_INNOVATION_PARTNER",
  ASSOCIATE = "ASSOCIATE_PARTNER",
  LOUNGE = "LOUNGE_PARTNER",
}

export enum StatePartnerPackage {
  PARTNER_STATE = "PARTNER_STATE",
  SUPPORTING_STATE_PARTNERSHIP = "SUPPORTING_STATE_PARTNER",
  KNOWLEDGE_PARTNER_STATE = "KNOWLEDGE_STATE_PARTNER",
}

export enum CountryPartnerPackage {
  PARTNER_COUNTRY = "PARTNER_COUNTRY",
  SUPPORTING_COUNTRY_PARTNERSHIP = "SUPPORTING_COUNTRY_PARTNER",
  KNOWLEDGE_COUNTRY_PARTNER = "KNOWLEDGE_COUNTRY_PARTNER",
}

export enum KnowledgeSessionsPackage {
  ROUNDTABLE = "ROUNDTABLE",
  FIRESIDE_CHAT = "FIRESIDE_CHAT",
  PANEL_DISCUSSION = "PANEL_DISCUSSION",
  MASTERCLASS_PODCAST = "MASTERCLASS",
}